uml: Update UML to linux 3.4.2 and tipcutils 2.0.2 (#2670)
UML environment was updated to linux 3.4.2, so that it builds on
Ubuntu 12.04. This version of Linux requires tipcutils to be updated
as well. Configuration files for both 32-bit and 64-bit are included.

The UML build script was also updated to be more robust in case of
failed downloads.

If you wish to use the old version of Linux and tipcutils, set the
following environment variables prior to building UML:

KVER=2.6.31
TIPCUTILSVER=1.0.4