avsv: set ALL SI flag when rec SUSI req (#2990)
Previously the SI ALL flag (in AVND_SU) was set in several places and its
purpose was not clear. For some cases the flag was not set causing problems
such as assert in avnd_su_si_oper_done()

Now the SI ALL flag is set in one location when the SUSI req is handled and
the purpose of the flag is documented.