immsv: Flag to allow ccb-error-string turned on just before callback (#3089)
Instead of opening up for saImmOiSetErrorString in the MDS thread
receiving the callback message from IMMND, the flag is turned on
by the application thread in the imma library just before the application
callback is invoked. This will avoid both having the flag on too early
allowing incorrect error string settings. It will also prevent the main
problem pointed out by the ticket, where the premature flag-on set by the
MDS thread is turned off by the termination of the previous callback in
the application thread. The previous callback may be wrapping up *after*
the MDS thread has processed the one or more subsequent callbacks.
The problem should only be relevant for cascading deletes.