IMM: Abort sync when deadlock detected between sync and PBE (#556)
The sync which is about to start is blocked waiting for ccbs to
complete. The logic that is periodically invoked to check this
is modified so that it detects if the PBE has been restarted AND
there are CCBs in critical. In that case deadlock is proclaimed
between PBE and sync and the sync is aborted.

There is always a waiting period of four seconds between syncs,
both after regular sync completion and aborted sync. This is to
allow the imm service to "come up for air", that is process other
urgent jobs, such as allowing a dump iteration to set epoch and start.

Testing this fix requires a testpatch that injects delay in the PBE
for the commit processing for a ccb, such as "ccb 3". Plus also
a delay in a restarted PBE to aquire the dump iterator.
This allows the tester tto trigger the ccb (using immcfg) and then
while the ccb is blocked, to kill the immnd coord, which will also
cause the PBE to terminate.

The restarted PBE will not be too quick in obtaining the dump
iterator, to allow the sync to get started => deadlock.

Without the fix, the deadlock will last for 20 seconds and then
both PBE and sync will abort and restart. With the patch, the sync
should get aborted soon after start. Then wait for 4 seconds before
restarting, allowing the PBE to proceed obtaining the dump and epoch.

Test patch will be provided on request.