immpbe: unlink any /tmp/imm.db.xxxx file and journal when exiting on error [#869]
Symptoms of the problem are dropped imm.db.xxxxxx files, mkstemp(3), and
imm.db.xxxxxx-journal files. The journal files are actually consistently dropped
as empty files even on successfull generation of imm.db.

The cause is simply an omission to unlink the temp file when exiting on error and
the fact that that sqlite never unlinks the journal file when journaling mode is
TRUNCATE. The journaling mode was altered for PBE a few years ago.

Solution is to catch all cases of controlled exit of PBE during the generation of
the temporary imm.db file and unlink these files before exiting. This ticket does
not take care of cleaning up files dropped by an uncontrolled crash of PBE.
To solve that problem safely and reliably we need to create all temporary imm.db
files in a sub directory to the configured IMMSV_PBE_TMP_DIR in immnd.conf. That
is a change of file representation that is more risky, needs more righorous sytem
testing and will thus be added as an enhancement.