amfd: create SI to store ckpt:ed dep_state [#926]
abort in standby amfd with back trace:

#0  0x00007f71264bcb35 in raise () from /lib64/libc.so.6
No symbol table info available.
#1  0x00007f71264be111 in abort () from /lib64/libc.so.6
No symbol table info available.
#2  0x00007f7127d51e07 in __osafassert_fail (__file=<optimized out>, __line=<optimized out>, __func=<optimized out>, __assertion=<optimized out>)
    at sysf_def.c:401
No locals.
#3  0x00000000004103e7 in dec_si_dep_state (cb=0x69c980, dec=<optimized out>) at avd_ckpt_dec.c:2042
        si_struct = 0x0

When an SI dependency is created on the active amfd it checkpoints the
dep_state. When this reach the standby amfd it has not yet created the SI so it
cannot store the state. The reason is the well-known race problem of standby
amfd being applier.

We fix this in standby amfd by creating the SI if it does not exist, this in
context of the check point handling.