imm: remove trace between fork and exec [#1077]
Any function/system call between fork and exec in the child process is a
potential risk of deadlock. The lock data structure is copied from the parent
process and a locked mutex will never get unlocked in the child since when the
parent process does unlock it won't affect the child copy of the lock.

To prevent this from happening, tracing is removed between fork and exec.