imm: In immnd_evt_proc_ccb_compl_rsp skip to end if PBE-OI is detached [#1096]
The case is where a CCB has passed validation by all involved OIs. Normally the
next step, if PBE is enabled, is to send the completed callbac to the PBE-OI.
The PBE actually commits the ccb/sqlite-transaction in the completed callback.

But in this case, the PBE has detached after the ccb has been built up, but
before imm-ram can send the completed-callback to the PBE. In this case it is
in princple ok for imm-ram to abort the CCB, since the PBE has not received
any completed callback and can not have comitted this ccb to disk.

But instead of complicating the ccb-state machine in ImmModel, just to optimize
an error case, the solution has been to just drop the send of completed to PBE
and let this error case merge with the case when the message was successfully
sent to the PBE yet the PBE restarted before replying.

The problem was that continued execution in this branch implied use of
the client node, which was missing.

The solution is to goto the end of the function.