amfd: do not allow duplicate addition of node in a nodegroup[#646]
This is version 2 based on changes pushed in #1142.

At present a node can be added in a nodegroup many times.

AMF is not checking if node is already present in saAmfNGNodeList of
nodegroup and hence not rejecting the ccb operation.

Patch adds to check to avoid duplicate addition of a node
in a nodegroup.