amfd: do not assign dependent csi if any sponsor csi is unassigned [#687]
A CSI is configured with CSI dependency on multiple sponsor CSIs. When
user UNLOCKs SUs, AMF is assigning the dependent CSI in the case when
only some sponsors are assigned.

Even if one sponsor is unassigned, AMF should not assign dependent CSI.
In this particular issue, one sponsor remained unassigned because
compcsType was not configured for the csType of unassigned sponsor.
At present, AMF does not have mechanism to strictly check the configuration
from assignment perspective.

AMF will not assign dependent CSI if any sponsor CSI is unassigned.