amfd: fix comp failover for 2N during shutdown op on ng [#1293]
In a SI dependency configuration if comp-failover occurs during shutdown operation
on ng for quiescing state, then SG does not become stable.

Although the context of the issue is NodeGroup, it is valid only for 2N model and
in the case when both active and assignments are part of the nodegroup.
SG remained unstable because standby assignments could not be removed.
This is due to the fact the standby SUSI pointer was not updated and remained NULL.


Patch ensures that AMFD finds the standby SUSI and remove is assignments.