amfd: validate ctcstype relationship while modifying comptype in comp [#1302]
Before modifying compType attribute in a comp, one more check on the validity of
SaAmfCtCsType is still required which means association (object of SaAmfCtCsType)
between this new comptype and cstype in SaAmfCompCsType (in case such an object
exists for this component) must also exists in the system. If ctcstype does not
exist then there will be problem in deciding components capability for a given
cstype. So reject the modify ccb if ctstype does not exist with any cstypes
supported by this component via compcstype.