amfd: assign SUs on unlocked nodes when locked NG is deleted [#1435].
A nodegroup is deleted in locked state when some of the nodes are kept in admin
unlocked state. Application remains in unassigned state.

In CCB apply for deletion of NG, AMF is not going for new assignments on the
nodes of NG. Since this is not documented and a valid use case, the problem
needs a fix.

Patch fixes the problem by calling the assignment logic in CCB apply for
deletion of NG.