amfd: fix comp term_failed state alarm [#1473]
In the reported problem when component enters into TERM_FAILED state
during clean up failure, AMFD raises alarm and clears it immediately.
In the reported problem saAmfNodeFailfastOnTerminationFailure is set false.

When component enters term_failed state AMFD raises alrams on it.
When all components are cleaned up, amfnd sends a su-failover request to amfd.
As a part of sufailover request amfd marks all the comps uninstantiated and
clears any pending alrams on them. Since saAmfNodeFailfastOnTerminationFailure is false,
clearence of alarm should be done as a part of repair operation on SU.
Even in the case when saAmfNodeFailfastOnTerminationFailure=1, Alarm should be cleared
only when AMFD has detected that node has gone for reboot.

Patch fixes both the problems.