amf: fix spec deviation of surestart escalation [#315]
During surestart recovery, currently AMF is not spec compliant in two
ways:
1)Individual component is terminated/cleaned up and then instantiated.
  Iinstead all componenets should be abruplty terminated and then instantiated
  honoring the instantiation level.
2)During termination/cleanup of individual component, saAmfDisableRestart flag is
  not honoured.
Ticket is updated with relevent spec sections for reference.

Patch solves the problem by cleaning up the component in the reverse order of
instantiation level and instantiation of compoment in the forward order. Also
before terminating a non-restartable component (saAmfDisableRestart=true),
assignments are switchover to another SU.