amfd: fix deletion of runtime objects of class SaAmfSIAssignment [#1540]
In the reported problem, application SI was lock and then unlocked.
When SI got successfully assigned, controller failover was done.
Upon successful fail-over, command : amf-state siass all | grep SI2
does not show any SUSI for this SI.

SUSI for the SI exists in AMFD data base, this can be verified by doing si-swap
operation on the SI or by dumping amfd state. When SI was locked, a job is added in
IMM job queue for the deletion of SUSI at active when it creates it in AMFD and at standby
AMFD it is added when it gets the checkpointing request from active. But when SI was
unlocked, only active added a job for susi creation, but in its checkpointing standby AMFD
was unable to add. The reason was in function check_to_create_immjob_at_standby_amfd(),
check was added for internal classIds of classes SaAmfSIAssignment amd SaAmfCSIAssignment
which is valid for deletion and update. But for creation of SUSI and CSICOMP in IMM name
of parent i.e SI and CSI is required. Because of this after failover when new
active clears its IMM job list, it only deletes the SUSI but since no create
job was added, user does not see SUSI and CSICOMP for SI and CSI,
respectively.

Patch resolves the problem by adding the check on SI and CSI class id also.