amfnd: do not run same clc commands at once [#1557] Currently if two errors happen at once, amfnd calls two clc cleanup commands in a row, and component fails into INSTANTIATION-FAILED state
As two clc cleanup commands for same component are called, that means two
flows of component life cycle are running at the same time. After the first
cleanup finishes, amfnd starts clc initiate command. For the reason mostly
due to timing, while amfnd is waiting for the return of clc initiate, amfnd
receives the return of second clc cleanup. Upon receiving the return of
second clc cleanup, the next step is that amfnd will start another clc
initiate if the max retry of clc initiate is large enough. But if the max
retry is 2, component fails into INSTANTIATION-FALED state, which should not
happen.

The cause of this issue is mainly because amfnd lets two clc commands
running concurrently. The second clc command is redundant while the first
one is in progress. Any errors happen during clc command execution, it
should be detected in the first one. The patch avoids the same clc commands
running in a row if amfnd has not got clc_resp (which indicates
timeout/error/ok) from the others.