amfd: adjust NPI app configuration during unlock-in op [#1564]
SG went to unstable for unlock operation on locked SG.

This issue is applicable for NPI apps only.
For a NPI SU, only one SI can be assigned to the whole SU. Remaining SIs if
configured will remain unassigned. Earliar creating a SU in the system was
allowed only in LOCK_IN state. So in typical steps, a user used to configure a SU
in the system in LOCK_IN state and after this user performs unlock-in operation
on the SU. During this UNLOCK_IN opearation, AMF adjusts SG configuration
attributes for example for NPI application it adjusts the value of
saAmfSGMaxActiveSIsperSU and saAmfSGMaxStandbySIsperSU. So issue will not be
reproducible in this way.
Now from 4.5 release creation of SU is allowed if either node or sg is in
LOCK_IN state. In this reported issue SUs are created by keeping SG in LOCK_IN
state.

Patch fixes the problem by adjusting the SG attributes during unlock-in
operation on SG, Node or NG.