cpsv: improve handling unlink and close non-collocated checkpoint [#1616] Problem: -------- There are several problems relating to closing and unlinking non-collocated checkpoint.
1. A non-collocated checkpoint is firstly created on SC-2. It is closed on SC-2. It is opened on PL-3.
It is unlinked. It is closes on PL-3. The replicas on SCs are not destroyed although the checkpoint is
unlinked and no client is using it.

2. A non-collocated checkpoint is firstly created on PL-3. It is closed on PL-3. It is opened on SC-2.
It is unlinked. It is closes on SC-2. The replicas on SCs and PL-3 are not destroyed although the
checkpoint is unlinked and no client is using it.

3. A non-collocated checkpoint is firstly created on PL-3. It is closed on PL-3. It is opened on PL-4.
It is unlinked. The replicas on SCs and PL-3 are destroyed although the checkpoint is using on PL-4.

Solution:
---------
The main cause of above problems is to use checking if non-collocated replica is on PL to decide
destroying the replicas. This mechanism is not correct in some cases. The solution is use another
mechanism which checks if there is any client using the checkpoint on the cluster by verifying if
the retention duration timer is active or not.

Test:
-----
Following test cases were executed for both non-collocated and collocated checkpoint to verify
the solution:
1. verify_unlink_ckpt_created_on_sc_before_close_it_from_sc
2. verify_unlink_ckpt_created_on_sc_before_close_it_from_pl
3. verify_unlink_ckpt_created_on_sc_after_close_it
4. verify_unlink_ckpt_created_on_pl_before_close_it_from_pl
5. verify_unlink_ckpt_created_on_pl_before_close_it_from_sc
6. verify_unlink_ckpt_created_on_pl_before_close_it_from_other_pl
7. verify_unlink_ckpt_created_on_pl_after_close_it