log: fix incorrect time output in log records [#1764]
If timestamp (`logTimeStamp`) is not provided by log client, log agent
will get the current time when getting the write log record request
and store the address which points to time holder variable in `logTimeStamp`.

But the time holder variable was an local variable in function `handle_log_record`,
then when exiting the function `handle_log_record`, the time holder variable's content will be deleted.
So, the original time data is unknown after the calling.

This patch fix the problem by not using the local variable to hold the time data.