smfd: choose the AU type with highest scope in merged procedure [#1810] When SMF_MERGE_TO_SINGLE_STEP feature is enabled and the step requires reboot, i.e. when saSmfBundleInstallOfflineScope or saSmfBundleRemoveOfflineScope is set to SA_SMF_CMD_SCOPE_PLM_EE and when smfSSAffectedNodesEnable is true then currently SMF uses the first occurrence of the AU's actedOn entry and throws the following error: - A software bundle requires reboot but the AU is a SU This patch changes this behaviour and chooses the AU/DU type that has the highest scope  i.e. safAmfNode scope for enabling node reboot. Therefore, If first entry is not of type safAmfNode, look for it in the rest of the actedOn list.  If not found in actedOn list, look in swAdd and swRemove bundles accordingly to choose the scope.
Testing Commands:
-----------------
Run a campaign with the merge feature enabled. The campaign should contain two procedures(into a single step)
with one procedure containining SUs in the actedOn field and another procedure with
plmExecEnv in the swAdd (or remove) field.

Testing, Expected Results:
--------------------------
With this patch, this error message should not occur - A software bundle requires reboot but the AU is a SU
and the campaign should succeed by proceeding to reboot based on the clustercontrollers attribute configuration accordingly.