cpnd: To erase element from section_db if inserting the element into local_section_db fails [#1843] Problem: -------- There are 2 steps when a section is added in to the database (function cpnd_ckpt_sec_add_db()): 1. The section information is inserted into section_db map 2. The section information is inserted into local_section_db map
In case the step 2. fails, the cpnd_ckpt_sec_add_db() returns a fault code immediately without
erasing the section information inserted in step 1. This leads a core dump triggered when
the checkpoint replica is deleted because there is an invalid element in the section_db map.

Solution:
---------
The solution is to erase the section information in step 1 in case the step 2 fails.