clm: Fix performance regression for standby SC startup time [#1845]
The OpenSAF start time for standby system controllers had degraded between
OpenSAF 4.7 and OpenSAF 5.0. The reason was that the CLM node director did not
reply to NID until the standby node had joined the cluster, and therefore the
whole startup of the standby controller was delayed. This was not a problem in
OpenSAF 4.7 because the CLM node agent was started much later in the NID
sequence. The CLM node agent has now been changed so that it replies to NID
properly during startup, which solves this regression.