amfnd: do not repair su without AMFD request in su-failover escalation [#1863]
In thre reported problem, lock operation on SU got timeout when a quiesced comp fault with
su-failover recovery.

AMFND calls avnd_err_su_repair() to repair the SU when su-failover recovery is going on.
When quiesced comp faults with su-failover recovery, AMFND launches cleaup of components.
In the meantime, AMFND gets removal of assignments and as a part of oper done it deletes SUSI
and callsavnd_err_su_repair(). Inside this function AMFND tries to instantiate UNINSTANTIATED comps.
No componnet is instantiated as they are in TERMINATING state. But SU_FAILOVER flag is reset inside
this function. Since AMFND clears the flag, it loses the context of su-failover escalation.
When first comp is cleaned up, AMFND instantiates it and thus the condition of all components
are terminated for informing AMFD about su-failover escalation is not met. Because of this
AMFD never responds for lock operation and it gets timed out.

As a part of fix AMFND does not call avnd_err_su_repair() during su-failover escalation and
also does not reset SU_FAILOVER flag for comp-failover recovery inside this function.