AMFND: Do not disable healthy SU [#1721]
Currently avnd_su_curr_info_del() is called in three places:

(1). su restart recovery

(2). su restart by admin op

(3). su is terminated by su_pres_msg

In case (1), (2), the code that reset SU's oper_state as DISABLED won't
be called. Only in (3), which lock-in SU (or node/ng) which SU is not
failed, that reset SU's oper_state as DISABLED. This will set local
variable @su->oper as DISABLED while SU OperationalState in amfd and
imm as ENABLED. This reset is not needed since if SU is healthy, its
oper state should be ENABLED. And this reset will cause SU won't be able
to recover after headless if there was a lock-in SU (node/ng) done before
headless.

Patch removes this reset of SU as DISABLED.