ntf: cluster rebooted with ntfd crashed on both controllers [#2006]
In AIS, it states "additionalText length must be consistent with lengthAdditionalText"
But NTFA did not check this. So, when data is passing to LOGA, ntfsv got invalid param.

This patch adds the check. Also fix few error in ntftest - used `sizeof` to calculate
the string length instead of `strlen`.

Besides, adding an early check to avoid allocating a too long `additionalText`.