AMF: Fix SG unstable from admin continuation of nodegroup after headless [#1987] V2
The SG becomes unstable because some variables used in nodegroup operation are not
restored after headless if this admin operation on nodegroup was interrupted just
before cluster goes into headless stage.

In order to restore nodegroup operation, AMF needs to know exactly whether nodegroup
operation was running during headless up on @susi assignment. If susi is in QUIESCED,
QUIESCING or being removed while its related entities su, si, sg are not in LOCKED
and SHUTTING_DOWN, that means either node or nodegroup MUST be in LOCKED or SHUTTING
DOWN. In case of SHUTTING_DOWN saAmfNGAdminState, that's enough to know a nodegroup
operation was running. However, if saAmfNGAdminState is in LOCKED, this case is an
ambiguity of locking a node. The reason of differentiation of locking a node or node
group is because 2N SG uses both AVD_SG_FSM_SG_ADMIN and AVD_SG_FSM_SU_OPER for node
group operation while AVD_SG_FSM_SU_OPER is only used for node operation. When 2N SG
uses AVD_SG_FSM_SG_ADMIN for nodegroup, the saAmfSGAdminState is borrowed (but not
updated to IMM) to run the admin operation sequence. Therefore, after headless if
AVD_SG_FSM_SG_ADMIN was being used for nodegroup then saAmfSGAdminState also needs to
be set.