amfnd: send recovery request to amfd for term-failed PI su [#2047] V2
A SU having one NPI and one PI comp moved to term-failed state during fresh assignments
and repair admin opreration does not work. It was NPI comp that faulted.

When SG is unlocked, AMFND initiates active assignments by instantiating the
NPI component and sending CSI set callback to PI comp. After instantiation failure
of NPI comp, AMFND tries to clean up the component. Cleanup fails.
AMFND marks comp and SU in TERM_FAILED state and terminates PI comp also. But AMFND
neither responds to AMFD for the completion of assignment nor it sends any recovery request.
Because of this SG remains unstable in REALIGN state. In this state, no admin
operation is allowed.

Patch solves the problem to send a recovery request to AMFD, so that it
deletes the assignments and marks SG stable.