imm: Convert syslog message to TRACE, to avoid filling up the syslog [#2136]
After a commit of an SMF campaign, the syslog was filled with thousands of log
messages with the text "Delete of PERSISTENT runtime object ...". Since the
syslog is a global log for all applications running on a node, OpenSAF ought to
be careful to not produce an excessive amount of syslog messages. The message
has therefore been converted to a trace message.